removed unused graphql queries and cleanup up file structure
