--- src/graphql/categories/category-mutations.js
@@ -1,10 +1,0 @@
-import gql from 'graphql-tag';
-
-export const ADD_CATEGORY = gql`
-  mutation addCategory($input: NewCategoryInput!) {
-    addCategory(input: $input) {
-      id
-      category
-    }
-  }
-`;
