+++ src/graphql/events/event-queries.js
@@ -1,40 +1,5 @@
 import gql from 'graphql-tag';
 
-export const ALL_EVENTS = gql`
-    query getAllEvents {
-        getAllEvents {
-            id
-            startTime
-            endTime
-            createDateTime
-            title
-            description
-            category_id
-            user_id
-            address
-            latitude
-            longitude
-            hashtags
-            modifiers
-            allergenWarnings
-            dietaryWarnings
-            photo
-            users {
-                id
-                email
-                firstName
-                lastName
-                longitude
-                latitude
-                status
-                address
-                gender
-                photo
-            }
-        }
-    }
-`;
-
 export const EVENT_BY_ID = gql`
     query Events($queryParams: EventInput) {
         Events(queryParams: $queryParams) {
